fixing edge cases
In some cases the state of the store could had change before react allows us to generate an unique id for the subscription.

due to this each subscription should revalidate their own state after the id was full committed and re-render if necessary.

If there is not change on the root state, the component will not re-render